Add files via upload
Due to confusion about the correct depth value in the dataset, the previous code used an incorrect measurement and gave uninterpretable results. This code uses the correct depth measurements and allows us to take subsets of the functional neurons and export the data as a csv file. It is necessary to do this in small sections due to the long run time and the internal server error Datajoint gives after a certain point.