Add files via upload
After we compiled all of the synapses into one dataframe, we finally plotted the graph using a Seaborn jointplot. This plot uses hexagons as bins for each point on the graph and also gives a density plot for the input and output depths. The lines added in are the estimations of the depths done by the MICrONS scientists.